+++ Assets/Scripts/Agent/Decider/DeciderXue.cs
@@ -87,7 +87,7 @@ public class DeciderXue : Decider
         var proposedAnticipationsSet = new HashSet<Anticipation>();
         foreach (var activatedInteraction in activatedInteractions)
         {
-            int proclivity = activatedInteraction.Weight * activatedInteraction.PostInteraction.Valence;
+            float proclivity = activatedInteraction.Weight * activatedInteraction.PostInteraction.Valence;
             var anticipation = new Anticipation(activatedInteraction.PostInteraction.Experiment, proclivity);
 
             if (!proposedAnticipationsSet.Contains(anticipation))
